Created rmsd_clustering module - Relocated functions - Cleaned up Redundant Code
